cast to long in sprintf to avoid warnings
mate sprintf quiet from glist_getcanvas

+++ src/g_template.c
@@ -1841,7 +1841,7 @@ static void plot_vis(t_gobj *z, t_glist *glist,
             color = fielddesc_getfloat(&x->x_outlinecolor, template, data, 1);
             if(color) 
             	sprintf(outline, "[::pdtk_canvas::get_color array_values "
-            		".x%lx]", glist_getcanvas(glist));
+            		".x%lx]", (long)glist_getcanvas(glist));
             else
             	numbertocolor(color, outline);
             if (wonset >= 0)
